make usage generator testable
